requirements.txt
# Core Streamlit application
streamlit>=1.28.0,<2.0.0

# Computer Vision and Image Processing
opencv-python>=4.8.0,<5.0.0
mediapipe==0.10.8
pillow>=9.0.0,<11.0.0
imutils>=0.5.4

# Machine Learning and AI (macOS with Apple Silicon)
tensorflow-macos==2.12.0
tensorflow-metal>=1.0.0
keras==2.12.0
deepface==0.0.79
numpy>=1.22.0,<1.24.0

# Data Processing and Analysis
pandas>=2.0.0,<3.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.17.0,<6.0.0

# System and Utilities
psutil>=5.9.0  # For system resource monitoring
requests>=2.31.0  # For HTTP requests (camera streams)
python-dotenv>=1.0.0  # For environment variables

# Development and Testing (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0

# Data Processing and Analysis
pandas>=2.0.0,<3.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.17.0,<6.0.0

# System and Utilities
psutil>=5.9.0  # For system resource monitoring
requests>=2.31.0  # For HTTP requests (camera streams)
python-dotenv>=1.0.0  # For environment variables

# Development and Testing (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
